refactor: (Headers.jsx): refactor in many components
Header refactor => Logo.jsx + NavigationLinks.jsx + PersonaHeartContainer.jsx with each SCSS
