+++ components/header/styles.module.scss
@@ -12,45 +12,3 @@
   padding: 0 120px;
   box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.25);
 }
-
-.navLinkContainer {
-  height: 100%;
-  display: flex;
-  align-items: center;
-  padding: 0 0.375rem;
-}
-
-.active {
-  font-weight: 600;
-  border-bottom: 3px solid $colorLink;
-  transition: all 0.3s ease-in-out;
-}
-
-.personaHeartContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 4.8125rem;
-}
-
-.userPastille {
-  display: flex;
-  background-color: $colorLink;
-  font-size: 0.8rem;
-  justify-content: center;
-  align-items: center;
-  height: 1.5rem;
-  width: 1.5rem;
-  border-radius: 50%;
-  color: white;
-  font-weight: bold;
-  &Black {
-    background-color: black;
-  }
-}
-
-.line {
-  height: 1.625rem;
-  width: 0.0625rem;
-  background-color: black;
-}
