+++ components/navigationLinks/styles.module.scss
@@ -1,0 +1,20 @@
+@import '/app/variables.module.scss';
+
+.navContainer {
+  display: flex;
+  justify-content: space-around;
+  width: 100%;
+  height: 100%;
+}
+
+.navLinkContainer {
+  display: flex;
+  align-items: center;
+  padding: 0 0.375rem;
+}
+
+.active {
+  font-weight: 600;
+  border-bottom: 3px solid $colorLink;
+  transition: all 0.3s ease-in-out;
+}
